Two changes in this commit - `generate_options` function moved into `cli.py` - `BuildOptions` moved into `options.py`
